ai_detection/requirements.txt
# AI-Generated Image Detection Pipeline
# Dependencies

numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
opencv-python-headless>=4.8.0
tqdm>=4.66.0
joblib>=1.3.0
PyWavelets>=1.4.0  # For PRNU sensor fingerprint extraction
ftfy
regex
flask
flask-cors
